requirements.txt
torch>=2.1
torchvision
diffusers>=0.27.0
transformers
safetensors
einops
Pillow
